Add client diversity exporter + test cases

+++ src/db/models/client_diversity.go
@@ -6,13 +6,13 @@ import (
 
 type ClientDiversity struct {
 	Timestamp  time.Time `json:"snapshot_timestamp"`
-	Prysm      int64     `json:"prysm"`
-	Lighthouse int64     `json:"lighthouse"`
-	Teku       int64     `json:"teku"`
-	Nimbus     int64     `json:"nimbus"`
-	Grandine   int64     `json:"grandine"`
-	Lodestar   int64     `json:"lodestar"`
-	Others     int64     `json:"others"`
+	Prysm      int       `json:"prysm"`
+	Lighthouse int       `json:"lighthouse"`
+	Teku       int       `json:"teku"`
+	Nimbus     int       `json:"nimbus"`
+	Grandine   int       `json:"grandine"`
+	Lodestar   int       `json:"lodestar"`
+	Others     int       `json:"others"`
 }
 
 func NewClientDiversity() ClientDiversity {
